+++ src/utils/exposeProperties.js
@@ -1,5 +1,9 @@
 'use strict';
 
 module.exports = function (obj, list = []) {
-	for (let k of list) Object.defineProperty(obj, k, { ...Object.getOwnPropertyDescriptor(obj, k), enumerable: true });
+	for (let k of list) Object.defineProperty(obj, k, {
+		...Object.getOwnPropertyDescriptor(obj, k),
+		enumerable: true,
+		configurable: false
+	});
 }